fix: resolved build issues feat: renamed handshake-utils to utils fix: refer the latest verion in the workspace for the dependant packages
