+++ handshake-nodejs-reporters/apps/test-wdio-cucumber/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "test-wdio-cucumber",
+	"name": "@handshake/test-wdio-cucumber",
 	"version": "1.0.0",
 	"private": true,
 	"type": "module",
@@ -20,6 +20,6 @@
 	"dependencies": {
 		"@wdio/cli": "^8.39.0",
 		"@wdio/mocha-framework": "8.39.0",
-		"wdio-handshake-reporter": "*"
+		"@handshake/wdio-handshake-reporter": "workspace:*"
 	}
 }
