+++ handshake-nodejs-reporters/package.json
@@ -1,6 +1,7 @@
 {
+	"name": "handshake-reporters",
 	"version": "0.4.2",
-	"packageManager": "npm@9.6.2",
+	"packageManager": "pnpm@8.15.6",
 	"type": "module",
 	"workspaces": [
 		"apps/*",
@@ -10,6 +11,7 @@
 	"scripts": {
 		"build": "turbo run build",
 		"dev": "turbo run dev",
+		"clean": "turbo clean && rimraf -g node_modules",
 		"sanity": "turbo run sanity --concurrency=1 --continue",
 		"test": "npm run build && turbo run test --concurrency=1 --continue",
 		"test-wdio": "turbo run test --continue --filter=test-wdio-*",
