+++ handshake-nodejs-reporters/packages/handshake-dashboard/src/pages/RUNS/index.tsx
@@ -1,6 +1,6 @@
 import React, { type ReactNode } from 'react';
 import Head from 'next/head';
-import { TEXT } from 'handshake-utils';
+import { TEXT } from '@handshake/utils';
 import { RunsPageContent } from 'components/about-test-runs/test-runs-page-layout';
 
 export default function AllTestRunsDisplayedHere(): ReactNode {
