+++ handshake-nodejs-reporters/packages/handshake-dashboard/src/pages/_document.tsx
@@ -1,4 +1,4 @@
-import { TEXT } from 'handshake-utils';
+import { TEXT } from '@handshake/utils';
 import { Html, Head, Main, NextScript } from 'next/document';
 // import Script from 'next/script';
 import type { ReactNode } from 'react';
