+++ handshake-nodejs-reporters/packages/handshake-dashboard/tsconfig.json
@@ -1,5 +1,5 @@
 {
-    "extends": "handshake-utils/tsconfig.ui.json",
+    "extends": "@handshake/utils/tsconfig.ui.json",
     "compilerOptions": {
         "baseUrl": "./src",
         "skipLibCheck": true
