+++ handshake-nodejs-reporters/packages/handshake-dashboard/tsconfig.node.json
@@ -1,3 +1,3 @@
 {
-    "extends": "handshake-utils/tsconfig.node-allow-jsx.json"
+    "extends": "@handshake/utils/tsconfig.node-allow-jsx.json"
 }
