+++ handshake-nodejs-reporters/packages/jest-handshake-reporter/package.json
@@ -29,7 +29,7 @@
 	"license": "MIT",
 	"dependencies": {
 		"@jest/reporters": "^29.7.0",
-		"common-handshakes": "workspace:*",
+		"@handshake/common-handshakes": "workspace:*",
 		"pino": "^9.2.0",
 		"superagent": "^9.0.2",
 		"typescript": "^5.5.2"
@@ -41,7 +41,7 @@
 		"@jest/globals": "^29.7.0",
 		"@types/async-lock": "^1.4.2",
 		"@types/jest": "^29.5.12",
-		"@types/node": "^20.14.7",
+		"@types/node": "^20.14.8",
 		"babel-jest": "^29.7.0",
 		"babel-plugin-transform-import-meta": "^2.2.1",
 		"jest": "^29.7.0"
