+++ handshake-nodejs-reporters/packages/jest-handshake-reporter/tsconfig.json
@@ -1,3 +1,3 @@
 {
-	"extends": "common-handshakes/node.json",
+	"extends": "@handshake/common-handshakes/node.json",
 }
