+++ handshake-nodejs-reporters/packages/utils/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "@handshake/handshake-utils",
+	"name": "@handshake/utils",
 	"version": "0.7.0",
 	"description": "util required for the handshake packages, internally used",
 	"main": "./dist/index.js",
