+++ handshake-nodejs-reporters/packages/utils/tsconfig.pure.node.json
@@ -12,7 +12,6 @@
 		"moduleResolution": "Node",
 		"module": "ESNext",
 		"lib": ["dom", "es2021"],
-		"types": ["node"],
 		"composite": false,
 		"forceConsistentCasingInFileNames": true
 	},
