+++ handshake-nodejs-reporters/packages/wdio-handshake-reporter/package.json
@@ -20,7 +20,7 @@
 	"dependencies": {
 		"@wdio/reporter": "^8.39.0",
 		"@wdio/types": "^8.39.0",
-		"common-handshakes": "workspace:*",
+		"@handshake/common-handshakes": "workspace:*",
 		"pino": "^9.2.0",
 		"superagent": "^9.0.2",
 		"typescript": "^5.5.2"
