+++ handshake-nodejs-reporters/packages/wdio-handshake-reporter/tsconfig.json
@@ -1,6 +1,3 @@
 {
-	"extends": "common-handshakes/node.json",
-	"compilerOptions": {
-		"types": ["@wdio/globals/types"]
-	},
+	"extends": "@handshake/common-handshakes/node.json"
 }
