+++ handshake-nodejs-reporters/pnpm-workspace.yaml
@@ -1,0 +1,3 @@
+packages:
+    - "apps/*"
+    - "packages/*"
